Accept proxy env variables from juju model config
