+++ .github/workflows/integration_test.yaml
@@ -8,4 +8,4 @@ jobs:
     uses: canonical/operator-workflows/.github/workflows/integration_test.yaml@main
     secrets: inherit
     with:
-      modules: '["test_charm", "test_default", "test_wsgi_path"]'
+      modules: '["test_charm", "test_proxy", "test_cos", "test_database", "test_default", "test_wsgi_path"]'
