+++ test_rock/app.py
@@ -95,6 +95,12 @@ def postgresql_status():
     return "FAIL"
 
 
+@app.route("/env")
+def get_env():
+    """Return environment variables"""
+    return jsonify(os.environ)
+
+
 app2 = Flask("app2")
 app2.config.from_prefixed_env()
 
@@ -102,4 +108,3 @@ app2.config.from_prefixed_env()
 @app2.route("/")
 def app2_hello_world():
     return "Hello, Many World!"
-
