+++ tests/unit/conftest.py
@@ -5,6 +5,7 @@
 
 # pylint: disable=protected-access,too-few-public-methods
 
+import os
 import typing
 
 import ops
@@ -117,3 +118,13 @@ def harness_fixture(monkeypatch) -> typing.Generator[Harness, None, None]:
 
     yield harness
     harness.cleanup()
+
+
+@pytest.fixture
+def set_environment_variable(request):
+    """Set the environment variables."""
+    environment_variables = request.param
+    os.environ.update(environment_variables)
+    yield
+    for environment_variable in environment_variables:
+        del os.environ[environment_variable]
